java datamodel - smaller memory footprint for FsContent, File, Directory

+++ bindings/java/src/org/sleuthkit/datamodel/ResultSetHelper.java
@@ -110,11 +110,11 @@ class ResultSetHelper {
 	 * @throws SQLException 
 	 */
 	File file(ResultSet rs, FileSystem fs) throws SQLException {
-		File f = new File(db, rs.getLong("obj_id"), rs.getLong("fs_obj_id"), rs.getLong("meta_addr"), rs.getInt("attr_type"),
-				rs.getInt("attr_id"), rs.getString("name"), rs.getLong("dir_type"),
-				rs.getLong("meta_type"), rs.getLong("dir_flags"), rs.getLong("meta_flags"), rs.getLong("size"),
-				rs.getLong("ctime"), rs.getLong("crtime"), rs.getLong("atime"), rs.getLong("mtime"),
-				rs.getLong("mode"), rs.getLong("uid"), rs.getLong("gid"), rs.getLong("known"), rs.getString("parent_path"), rs.getString("md5"));
+		File f = new File(db, rs.getLong("obj_id"), rs.getLong("fs_obj_id"), rs.getLong("meta_addr"), rs.getShort("attr_type"),
+				rs.getShort("attr_id"), rs.getString("name"), rs.getShort("dir_type"),
+				rs.getShort("meta_type"), rs.getByte("dir_flags"), rs.getByte("meta_flags"), rs.getLong("size"),
+				rs.getInt("ctime"), rs.getInt("crtime"), rs.getInt("atime"), rs.getInt("mtime"),
+				rs.getShort("mode"), rs.getInt("uid"), rs.getInt("gid"), rs.getByte("known"), rs.getString("parent_path"), rs.getString("md5"));
 		f.setFileSystem(fs);
 		return f;
 	}
@@ -128,11 +128,11 @@ class ResultSetHelper {
 	 * @throws SQLException thrown if SQL error occurred
 	 */
 	Directory directory(ResultSet rs, FileSystem fs, String name) throws SQLException {
-		Directory dir = new Directory(db, rs.getLong("obj_id"), rs.getLong("fs_obj_id"), rs.getLong("meta_addr"), rs.getInt("attr_type"),
-				rs.getInt("attr_id"), name, rs.getLong("dir_type"),
-				rs.getLong("meta_type"), rs.getLong("dir_flags"), rs.getLong("meta_flags"), rs.getLong("size"),
-				rs.getLong("ctime"), rs.getLong("crtime"), rs.getLong("atime"), rs.getLong("mtime"),
-				rs.getLong("mode"), rs.getLong("uid"), rs.getLong("gid"), rs.getLong("known"), rs.getString("parent_path"), rs.getString("md5"));
+		Directory dir = new Directory(db, rs.getLong("obj_id"), rs.getLong("fs_obj_id"), rs.getLong("meta_addr"), rs.getShort("attr_type"),
+				rs.getShort("attr_id"), name, rs.getShort("dir_type"),
+				rs.getShort("meta_type"), rs.getByte("dir_flags"), rs.getByte("meta_flags"), rs.getLong("size"),
+				rs.getInt("ctime"), rs.getInt("crtime"), rs.getInt("atime"), rs.getInt("mtime"),
+				rs.getShort("mode"), rs.getInt("uid"), rs.getInt("gid"), rs.getByte("known"), rs.getString("parent_path"), rs.getString("md5"));
 		dir.setFileSystem(fs);
 		return dir;
 	}
